python/pyproject.toml
[tool.poetry]
name = "blackbox-python-sdk"
version = "0.1.0"
description = "Automatic evolution tracking for AI black box functions"
authors = ["Blackbox Contributors"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/ravibhim/blackbox-clients"
keywords = ["ai", "llm", "tracing", "observability", "blackbox"]
packages = [{include = "blackbox_python_sdk", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
httpx = "^0.28.1"
typing-extensions = "^4.0"
pydantic = "^2.11.0"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-instrumentation-openai = "^0.49.0"
opentelemetry-instrumentation-anthropic = "^0.49.0"
openinference-instrumentation-groq = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^0.24.0"
black = "^25.1.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
